Added intervals and changed ai request for project
The function make 20 moves needs to be changed to instead startGame() and the button should be relabeled as well. The function make move now goes in a for loop that iterates 9 times to ensure that each player moves all of his/hers pieces in one turn. Added the changes to the AI post request.